Add directory for Hephaestus source terms
