+++ src/sources/sources.hpp
@@ -1,0 +1,17 @@
+#pragma once
+#include "div_free_source.hpp"
+#include "scalar_potential_source.hpp"
+
+namespace hephaestus {
+
+class Sources : public mfem::NamedFieldsMap<hephaestus::Source> {
+public:
+  void Init(mfem::NamedFieldsMap<mfem::ParGridFunction> &variables,
+            const mfem::NamedFieldsMap<mfem::ParFiniteElementSpace> &fespaces,
+            hephaestus::BCMap &bc_map,
+            hephaestus::DomainProperties &domain_properties);
+  void Apply(mfem::ParLinearForm *lf);
+  void SubtractSources(mfem::ParGridFunction *gf);
+};
+
+}; // namespace hephaestus
